refactor: decompose Home.tsx + migrate icons to Lucide
Home.tsx shrinks 4,132 -> 1,465 lines with zero behavioral change, proven by
token-level parity checks against the previous revision at every stage:
usePendingRequests hook owns the 13 pending-request states and the socket
mapping effects (identical API, subscription untouched); the render splits
into HomeProgress/HomeIdle/HomeRequests/HomeModals with Home remaining the
sole state owner; 15 handler bodies relocate verbatim into action modules
behind a context whose property names equal the original closure variables;
pure input parsing extracted with tests. One verified consolidation: the
vault-signing ingestion block that existed as two verbatim copies is now
ingestVaultSigningRequest.

Icons: the Feather + MaterialCommunityIcons mix migrates to lucide-react-native
(pure JS on the existing react-native-svg; exact-pinned). All sizes, colors and
accessibility props preserved. react-native-vector-icons stays as a dependency
until a device-verified native rebuild removes its fonts.

Device pass before release should exercise: socket delivery of every request
type, pull-to-refresh vault-sign ingestion, pairing + batch sync end-to-end,
approve and reject paths, QR + manual input parsing, nonce replenish.
