+++ package.json
@@ -61,6 +61,7 @@
     "https-browserify": "1.0.0",
     "i18next": "26.3.1",
     "jail-monkey": "3.0.0",
+    "lucide-react-native": "1.24.0",
     "postinstall-postinstall": "2.1.0",
     "process": "0.11.10",
     "react": "19.2.3",
