+++ src/components/TxSent/TxSent.tsx
@@ -7,7 +7,7 @@ import {
   ScrollView,
   Linking,
 } from 'react-native';
-import Icon from 'react-native-vector-icons/Feather';
+import { CircleCheck } from 'lucide-react-native';
 import { useTranslation } from 'react-i18next';
 import { useTheme } from '../../hooks';
 import { explorerTxUrl } from '../../lib/explorerUrl';
@@ -60,7 +60,7 @@ const TxSent = (props: {
               Layout.alignItemsCenter,
             ]}
           >
-            <Icon name="check-circle" size={60} color={Colors.textGray400} />
+            <CircleCheck size={60} color={Colors.textGray400} />
             <Text
               style={[
                 Fonts.textBold,
