+++ src/screens/Create/Create.tsx
@@ -12,7 +12,7 @@ import {
   Linking,
 } from 'react-native';
 import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view';
-import Icon from 'react-native-vector-icons/Feather';
+import { ChevronLeft, Eye, EyeOff } from 'lucide-react-native';
 import Toast from 'react-native-toast-message';
 import { useTranslation } from 'react-i18next';
 import { useTheme } from '../../hooks';
@@ -334,7 +334,7 @@ function Create({ navigation }: Props) {
           onPress={() => navigation.navigate('Welcome')}
           style={[Layout.row]}
         >
-          <Icon name="chevron-left" size={20} color={Colors.primary} />
+          <ChevronLeft size={20} color={Colors.primary} />
           <Text
             style={[Fonts.textSmall, Fonts.textPrimary, Gutters.tinyLPadding]}
           >
@@ -418,7 +418,11 @@ function Create({ navigation }: Props) {
               onPress={handlePasswordVisibility}
               style={Common.inputIcon}
             >
-              <Icon name={rightIcon} size={20} color={Colors.primary} />
+              {rightIcon === 'eye' ? (
+                <Eye size={20} color={Colors.primary} />
+              ) : (
+                <EyeOff size={20} color={Colors.primary} />
+              )}
             </TouchableOpacity>
           </View>
           <View
@@ -446,7 +450,11 @@ function Create({ navigation }: Props) {
               onPress={handlePasswordVisibilityConfirm}
               style={Common.inputIcon}
             >
-              <Icon name={rightIconConfirm} size={20} color={Colors.primary} />
+              {rightIconConfirm === 'eye' ? (
+                <Eye size={20} color={Colors.primary} />
+              ) : (
+                <EyeOff size={20} color={Colors.primary} />
+              )}
             </TouchableOpacity>
           </View>
           <Text
