+++ src/screens/LavaMoatTest/LavaMoatTest.tsx
@@ -8,7 +8,7 @@ import {
   StyleSheet,
   SafeAreaView,
 } from 'react-native';
-import Icon from 'react-native-vector-icons/Feather';
+import { ArrowLeft } from 'lucide-react-native';
 import { useTheme } from '../../hooks';
 
 interface TestResult {
@@ -421,7 +421,7 @@ const LavaMoatTest: React.FC<Props> = ({ navigation }) => {
             onPress={() => navigation.goBack()}
             style={styles.backButton}
           >
-            <Icon name="arrow-left" size={24} color="#fff" />
+            <ArrowLeft size={24} color="#fff" />
           </TouchableOpacity>
           <Text style={styles.title}>🔒 LavaMoat Security Test</Text>
           <View style={{ width: 24 }} />
